pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.127.1",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.2",
    "pydantic-settings>=2.12.0",
    "pymysql>=1.1.2",
    "ruff>=0.14.10",
    "sqlalchemy>=2.0.45",
]
